social_connect_app_backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-cors-headers==3.13.0
setuptools==65.5.1
pytz==2022.6
pip==22.3.1
psycopg2==2.8.4
SQLAlchemy==1.4.44
importlib-metadata==5.0.0